requirements.txt
requests
python-dotenv
flask
opencv-python
matplotlib
numpy
